% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_missdiva_loaded}
\alias{check_missdiva_loaded}
\title{Check that missdiva package has loaded successfully}
\usage{
check_missdiva_loaded()
}
\description{
Simple function to check that missdiva package has loaded
  successfully. Prints "missdiva loaded successfully!" if so.
}
