% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{coverage_plots}
\alias{coverage_plots}
\title{Make exploratory plots of missingness by position and sample}
\usage{
coverage_plots(coverage, threshold)
}
\arguments{
\item{coverage}{coverage matrix.}

\item{threshold}{below this threshold is coded as missing.}
}
\description{
Make exploratory plots of missingness by position and sample
}
