% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_coverage_matrix}
\alias{get_coverage_matrix}
\title{Extract coverage matrix from vcf object}
\usage{
get_coverage_matrix(file = NULL, vcf = NULL, verbose = TRUE)
}
\arguments{
\item{file}{path to vcf file.}

\item{vcf}{object of class \code{vcfR}.}

\item{verbose}{if reading from file, whether to read in verbose manner.}
}
\description{
Extract coverage matrix from vcf object.
}
