% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_missing_distance}
\alias{get_missing_distance}
\title{Get pairwise distances based on concordance}
\usage{
get_missing_distance(missing_matrix = NULL)
}
\arguments{
\item{missing_matrix}{matrix of missingness encoded as a binary variable}
}
\description{
Get pairwise distances based on concordance.
}
