% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_pcoa}
\alias{get_pcoa}
\title{Principal coordinates analysis (PCoA) on distance matrix}
\usage{
get_pcoa(distance_matrix = NULL)
}
\arguments{
\item{distance_matrix}{matrix of distances.}
}
\description{
Principal coordinates analysis (PCoA) on distance matrix.
}
