% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pca_missingness_plot}
\alias{pca_missingness_plot}
\title{Do a PCA plot of missingness}
\usage{
pca_missingness_plot(pca_dist = NULL, title = NULL)
}
\arguments{
\item{pca_distance}{PCA on distance matrix}
}
\description{
plot the results of PCA using missingness distance matrix
}
