% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage_matrix}
\alias{plot_coverage_matrix}
\title{Plot matrix of coverage for all samples and all loci}
\usage{
plot_coverage_matrix(x = NULL)
}
\arguments{
\item{x}{matrix of coverage.}
}
\description{
Plot matrix of coverage for all samples and all loci.
}
