% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_locus_quality_hist}
\alias{plot_locus_quality_hist}
\title{Plot histogram of quality of each locus}
\usage{
plot_locus_quality_hist(missing_matrix)
}
\arguments{
\item{missing_matrix}{matrix of missingness encoded as a binary variable.}
}
\description{
Plot histogram of quality of each locus.
}
