% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pcoa}
\alias{plot_pcoa}
\title{PCoA plot}
\usage{
plot_pcoa(pcoa = NULL, title = NULL)
}
\arguments{
\item{pcoa}{result of running PCoA, e.g. through the function
\code{get_pcoa()}.}

\item{title}{plot title.}
}
\description{
Plot the first two PCoA axes.
}
