% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_quality}
\alias{plot_sample_quality}
\title{Plot quality of each sample}
\usage{
plot_sample_quality(missing_matrix)
}
\arguments{
\item{missing_matrix}{matrix of missingness encoded as a binary variable.}
}
\description{
Plot quality of each sample.
}
