% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_quality_hist}
\alias{plot_sample_quality_hist}
\title{Plot histogram of quality of each sample}
\usage{
plot_sample_quality_hist(missing_matrix)
}
\arguments{
\item{missing_matrix}{matrix of missingness encoded as a binary variable.}
}
\description{
Plot histogram of quality of each sample.
}
