% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in vcf object}
\usage{
read_vcf(file, verbose = TRUE)
}
\arguments{
\item{file}{path to vcf file.}

\item{verbose}{if reading from file, whether to read in verbose manner.}
}
\description{
Read in vcf object.
}
