% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vcf_concat_crude}
\alias{vcf_concat_crude}
\title{Crude method for concatenating vcfs}
\usage{
vcf_concat_crude(input_paths, output_path, check_overwrite = TRUE)
}
\arguments{
\item{input_paths}{vector of file paths to input vcf files.}

\item{output_path}{single file path specifying thr output vcf file.}

\item{check_overwrite}{Boolean, whether to prompt user before overwriting
existing files.}
}
\description{
A vector of input vcf file paths is specified. Each file is read
  in, and any lines starting with the "#" symbol are taken to be header
  lines. Header lines are retained for the first input file, but dropped
  thereafter when concatenating files. The final concatenated data is printed
  to file.
}
