% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coverage_plots}
\alias{coverage_plots}
\title{Making exploratory plots of missingness by position and sample}
\usage{
coverage_plots(coverage, threshold)
}
\arguments{
\item{missing_matrix}{matrix of missingness encoded as a binary variable}
}
\description{
from the binary missingness matrix compute and plot missingness
}
