% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_missing_matrix}
\alias{make_missing_matrix}
\title{Make a matrix of missingness using coverage matrix}
\usage{
make_missing_matrix(coverage = NULL, threshold = NULL)
}
\arguments{
\item{coverage}{coverage matrix, should be numeric values, can contain NAs.}

\item{threshold}{numeric object saying what the minimum coverage threshold is.}
}
\description{
Use coverage values extracted by the coverage function and using a variable threshold make a matrix of missingness, with a binary variable.
}
